number of rays calculate
	modified:   TowerDefenceANTS_Prototype/bin/controllers/basics/ANTSSimpleRayLightController.class
	modified:   TowerDefenceANTS_Prototype/bin/model/basics/ANTSSimpleRayLightModel.class
	modified:   TowerDefenceANTS_Prototype/bin/model/basics/ANTSSimpleSourceLightModel.class
	modified:   TowerDefenceANTS_Prototype/bin/view/basics/ANTSSimpleRayLightView.class
	modified:   TowerDefenceANTS_Prototype/bin/view/basics/ANTSSimpleSourceLightView.class
	modified:   TowerDefenceANTS_Prototype/src/controllers/basics/ANTSSimpleRayLightController.java
	modified:   TowerDefenceANTS_Prototype/src/model/basics/ANTSSimpleRayLightModel.java
	modified:   TowerDefenceANTS_Prototype/src/model/basics/ANTSSimpleSourceLightModel.java
	modified:   TowerDefenceANTS_Prototype/src/view/basics/ANTSSimpleRayLightView.java
	modified:   TowerDefenceANTS_Prototype/src/view/basics/ANTSSimpleSourceLightView.java

+++ TowerDefenceANTS_Prototype/src/controllers/basics/ANTSSimpleRayLightController.java
@@ -25,7 +25,7 @@ public class ANTSSimpleRayLightController implements ANTSIController
 	private ANTSSimpleRayLightModel model;
 	
 	
-	public ANTSSimpleRayLightController(ANTSSimpleSourceLightModel lightModel, int startAngle)
+	public ANTSSimpleRayLightController(ANTSSimpleSourceLightModel lightModel, double startAngle)
 	{
 		this.model = new ANTSSimpleRayLightModel(lightModel,startAngle);
 		this.view = new ANTSSimpleRayLightView(this.model);
